+++ Sources/Swift/Integrations/SessionReplay/SentryOnDemandReplay.swift
@@ -1,4 +1,4 @@
-// swiftlint:disable file_length
+// swiftlint:disable file_length type_body_length
 #if canImport(UIKit) && !SENTRY_NO_UIKIT
 #if os(iOS) || os(tvOS)
 
@@ -16,7 +16,8 @@ class SentryOnDemandReplay: NSObject, SentryReplayVideoMaker {
     private let _outputPath: String
     private var _totalFrames = 0
     private let dateProvider: SentryCurrentDateProvider
-    private let workingQueue: SentryDispatchQueueWrapper
+    private let processingQueue: SentryDispatchQueueWrapper
+    private let assetWorkerQueue: SentryDispatchQueueWrapper
     private var _frames = [SentryReplayFrame]()
 
     #if SENTRY_TEST || SENTRY_TEST_CI || DEBUG
@@ -31,14 +32,31 @@ class SentryOnDemandReplay: NSObject, SentryReplayVideoMaker {
     var frameRate = 1
     var cacheMaxSize = UInt.max
         
-    init(outputPath: String, workingQueue: SentryDispatchQueueWrapper, dateProvider: SentryCurrentDateProvider) {
+    init(
+        outputPath: String,
+        processingQueue: SentryDispatchQueueWrapper,
+        assetWorkerQueue: SentryDispatchQueueWrapper,
+        dateProvider: SentryCurrentDateProvider
+    ) {
+        assert(processingQueue != assetWorkerQueue, "Processing and asset worker queue must not be the same.")
         self._outputPath = outputPath
         self.dateProvider = dateProvider
-        self.workingQueue = workingQueue
+        self.processingQueue = processingQueue
+        self.assetWorkerQueue = assetWorkerQueue
     }
         
-    convenience init(withContentFrom outputPath: String, workingQueue: SentryDispatchQueueWrapper, dateProvider: SentryCurrentDateProvider) {
-        self.init(outputPath: outputPath, workingQueue: workingQueue, dateProvider: dateProvider)
+    convenience init(
+        withContentFrom outputPath: String,
+        processingQueue: SentryDispatchQueueWrapper,
+        assetWorkerQueue: SentryDispatchQueueWrapper,
+        dateProvider: SentryCurrentDateProvider
+    ) {
+        self.init(
+            outputPath: outputPath,
+            processingQueue: processingQueue,
+            assetWorkerQueue: assetWorkerQueue,
+            dateProvider: dateProvider
+        )
         loadFrames(fromPath: outputPath)
     }
 
@@ -57,42 +75,34 @@ class SentryOnDemandReplay: NSObject, SentryReplayVideoMaker {
             }.sorted { $0.time < $1.time }
             SentryLog.debug("[Session Replay] Loaded \(content.count) files into \(_frames.count) frames from path: \(path)")
         } catch {
-            SentryLog.error("[Session Replay] Could not list frames from replay: \(error.localizedDescription)")
+            SentryLog.error("[Session Replay] Could not list frames from replay, reason: \(error.localizedDescription)")
         }
     }
 
-    convenience init(outputPath: String) {
-        self.init(outputPath: outputPath,
-                  workingQueue: SentryDispatchQueueWrapper(name: "io.sentry.onDemandReplay", attributes: nil),
-                  dateProvider: SentryDefaultCurrentDateProvider())
-    }
-    
-    convenience init(withContentFrom outputPath: String) {
-        self.init(withContentFrom: outputPath,
-                  workingQueue: SentryDispatchQueueWrapper(name: "io.sentry.onDemandReplay", attributes: nil),
-                  dateProvider: SentryDefaultCurrentDateProvider())
-    }
-
     func addFrameAsync(image: UIImage, forScreen: String?) {
-        workingQueue.dispatchAsync({
+        SentryLog.debug("[Session Replay] Adding frame async for screen: \(forScreen ?? "nil")")
+        // Dispatch the frame addition to a background queue to avoid blocking the main queue.
+        // This must be on the processing queue to avoid deadlocks.
+        processingQueue.dispatchAsync {
             self.addFrame(image: image, forScreen: forScreen)
-        })
+        }
     }
     
     private func addFrame(image: UIImage, forScreen: String?) {
         SentryLog.debug("[Session Replay] Adding frame to replay, screen: \(forScreen ?? "nil")")
-        guard let data = rescaleImage(image)?.pngData() else { 
+        guard let data = rescaleImage(image)?.pngData() else {
             SentryLog.error("[Session Replay] Could not rescale image, dropping frame")
             return
         }
-        
+
         let date = dateProvider.date()
         let imagePath = (_outputPath as NSString).appendingPathComponent("\(date.timeIntervalSinceReferenceDate).png")
         do {
-            SentryLog.debug("[Session Replay] Saving frame to path: \(imagePath)")
-            try data.write(to: URL(fileURLWithPath: imagePath))
+            let url = URL(fileURLWithPath: imagePath)
+            SentryLog.debug("[Session Replay] Saving frame to file URL: \(url)")
+            try data.write(to: url)
         } catch {
-            SentryLog.error("[Session Replay] Could not save replay frame. Error: \(error)")
+            SentryLog.error("[Session Replay] Could not save replay frame, reason: \(error)")
             return
         }
         _frames.append(SentryReplayFrame(imagePath: imagePath, time: date, screenName: forScreen))
@@ -100,11 +110,12 @@ class SentryOnDemandReplay: NSObject, SentryReplayVideoMaker {
         // Remove the oldest frames if the cache size exceeds the maximum size.
         while _frames.count > cacheMaxSize {
             let first = _frames.removeFirst()
-            SentryLog.debug("[Session Replay] Removing oldest frame at path: \(first.imagePath)")
-            try? FileManager.default.removeItem(at: URL(fileURLWithPath: first.imagePath))
+            let url = URL(fileURLWithPath: first.imagePath)
+            SentryLog.debug("[Session Replay] Removing oldest frame at file URL: \(url.path)")
+            try? FileManager.default.removeItem(at: url)
         }
         _totalFrames += 1
-        SentryLog.debug("[Session Replay] Increased total frames to: \(_totalFrames)")
+        SentryLog.debug("[Session Replay] Added frame, total frames counter: \(_totalFrames), current frames count: \(_frames.count)")
     }
     
     private func rescaleImage(_ originalImage: UIImage) -> UIImage? {
@@ -122,7 +133,7 @@ class SentryOnDemandReplay: NSObject, SentryReplayVideoMaker {
     }
     
     func releaseFramesUntil(_ date: Date) {
-        workingQueue.dispatchAsync ({
+        processingQueue.dispatchAsync {
             SentryLog.debug("[Session Replay] Releasing frames until date: \(date)")
             while let first = self._frames.first, first.time < date {
                 self._frames.removeFirst()
@@ -135,164 +146,264 @@ class SentryOnDemandReplay: NSObject, SentryReplayVideoMaker {
                 }
             }
             SentryLog.debug("[Session Replay] Frames released, remaining frames count: \(self._frames.count)")
-        })
+        }
     }
         
     var oldestFrameDate: Date? {
         return _frames.first?.time
     }
-    
-    func createVideoWith(beginning: Date, end: Date) throws -> [SentryVideoInfo] {
-        SentryLog.debug("[Session Replay] Creating video with beginning: \(beginning), end: \(end)")
-        let videoFrames = filterFrames(beginning: beginning, end: end)
-        var frameCount = 0
-        
-        var videos = [SentryVideoInfo]()
-        
-        while frameCount < videoFrames.count {
-            let outputFileURL = URL(fileURLWithPath: _outputPath.appending("/\(videoFrames[frameCount].time.timeIntervalSinceReferenceDate).mp4"))
-            SentryLog.debug("[Session Replay] Rendering video with output file URL: \(outputFileURL)")
-            if let videoInfo = try renderVideo(with: videoFrames, from: &frameCount, at: outputFileURL) {
-                videos.append(videoInfo)
-            } else {
-                frameCount++
-            }
+
+    func createVideoInBackgroundWith(beginning: Date, end: Date, completion: @escaping ([SentryVideoInfo]) -> Void) {
+        // Note: In Swift it is best practice to use `Result<Value, Error>` instead of `(Value?, Error?)`
+        //       Due to interoperability with Objective-C and @objc, we can not use Result for the completion callback.
+        SentryLog.debug("[Session Replay] Creating video in background with beginning: \(beginning), end: \(end)")
+        processingQueue.dispatchAsync {
+            let videos = self.createVideoWith(beginning: beginning, end: end)
+            SentryLog.debug("[Session Replay] Finished creating video in background with \(videos.count) segments")
+            completion(videos)
         }
+    }
+
+    func createVideoWith(beginning: Date, end: Date) -> [SentryVideoInfo] {
+        SentryLog.debug("[Session Replay] Creating video with beginning: \(beginning), end: \(end)")
+        // Note: In previous implementations this method was wrapped by a sync call to the processing queue.
+        // As this method is already called from the processing queue, we must remove the sync call.
+        let videoFrames = self._frames.filter { $0.time >= beginning && $0.time <= end }
+        var frameCount = 0
+
+        var videos = [SentryVideoInfo]()
+
+        while frameCount < videoFrames.count {
+            let frame = videoFrames[frameCount]
+            let outputFileURL = URL(fileURLWithPath: _outputPath)
+                .appendingPathComponent("\(frame.time.timeIntervalSinceReferenceDate)")
+                .appendingPathExtension("mp4")
+
+            let group = DispatchGroup()
+            var currentError: Error?
+
+            group.enter()
+            self.renderVideo(with: videoFrames, from: frameCount, at: outputFileURL) { result in
+                switch result {
+                case .success(let videoResult):
+                    // Set the frame count/offset to the new index that is returned by the completion block.
+                    // This is important to avoid processing the same frame multiple times.
+                    frameCount = videoResult.finalFrameIndex
+                    SentryLog.debug("[Session Replay] Finished rendering video, frame count moved to: \(frameCount)")
+
+                    // Append the video info to the videos array.
+                    // In case no video info is returned, skip the segment.
+                    if let videoInfo = videoResult.info {
+                        videos.append(videoInfo)
+                    }
+                case .failure(let error):
+                    SentryLog.error("[Session Replay] Failed to render video with error: \(error)")
+                    currentError = error
+                }
+                group.leave()
+            }
+
+            // Calling group.wait will block the `processingQueue` until the video rendering completes or a timeout occurs.
+            // It is imporant that the renderVideo completion block signals the group.
+            // The queue used by render video must have a higher priority than the processing queue to reduce thread inversion.
+            // Otherwise, it could lead to queue starvation and a deadlock/timeout.
+            guard group.wait(timeout: .now() + 10) == .success else {
+                SentryLog.error("[Session Replay] Timeout while waiting for video rendering to finish, returning \(videos.count) videos")
+                return videos
+            }
+
+            // If there was an error, log it and exit the loop.
+            if let error = currentError {
+                // Until v8.50.2 the error was propagated to the completion block, discarding any generated video.
+                // Instead this will "silently" fail by only logging the error and returning the successfully generated videos.
+                SentryLog.error("[Session Replay] Error while rendering video: \(error), returning \(videos.count) videos")
+                return videos
+            }
+
+            SentryLog.debug("[Session Replay] Finished rendering video, frame count moved to: \(frameCount)")
+        }
+
+        SentryLog.debug("[Session Replay] Finished creating video with \(videos.count) segments")
         return videos
     }
 
-    // swiftlint:disable:next function_body_length
-    private func renderVideo(with videoFrames: [SentryReplayFrame], from: inout Int, at outputFileURL: URL) throws -> SentryVideoInfo? {
-        SentryLog.debug("[Session Replay] Rendering video with \(videoFrames.count) video frames, from index: \(from), output file URL: \(outputFileURL)")
-        guard from < videoFrames.count, let image = UIImage(contentsOfFile: videoFrames[from].imagePath) else { 
-            SentryLog.error("[Session Replay] Could not render video, reason: frame not found")
-            return nil 
+    // swiftlint:disable function_body_length cyclomatic_complexity
+    private func renderVideo(with videoFrames: [SentryReplayFrame], from: Int, at outputFileURL: URL, completion: @escaping (Result<SentryRenderVideoResult, Error>) -> Void) {
+        SentryLog.debug("[Session Replay] Rendering video with \(videoFrames.count) frames, from index: \(from), to output url: \(outputFileURL)")
+        guard from < videoFrames.count else {
+            SentryLog.error("[Session Replay] Failed to render video, reason: index out of bounds")
+            return completion(.success(SentryRenderVideoResult(
+                info: nil,
+                finalFrameIndex: from
+            )))
         }
+        guard let image = UIImage(contentsOfFile: videoFrames[from].imagePath) else {
+            SentryLog.error("[Session Replay] Failed to render video, reason: can't read image at path: \(videoFrames[from].imagePath)")
+            return completion(.success(SentryRenderVideoResult(
+                info: nil,
+                finalFrameIndex: from
+            )))
+        }
+        
         let videoWidth = image.size.width * CGFloat(videoScale)
         let videoHeight = image.size.height * CGFloat(videoScale)
-        
-        let videoWriter = try AVAssetWriter(url: outputFileURL, fileType: .mp4)
-        let videoWriterInput = AVAssetWriterInput(mediaType: .video, outputSettings: createVideoSettings(width: videoWidth, height: videoHeight))
-        
-        guard let currentPixelBuffer = SentryPixelBuffer(size: CGSize(width: videoWidth, height: videoHeight), videoWriterInput: videoWriterInput) else {
-            SentryLog.error("[Session Replay] Failed to render video, reason: pixel buffer creation failed")
-            throw SentryOnDemandReplayError.cantCreatePixelBuffer
+        let pixelSize = CGSize(width: videoWidth, height: videoHeight)
+
+        SentryLog.debug("[Session Replay] Creating video writer with output file URL: \(outputFileURL)")
+        let videoWriter: AVAssetWriter
+        do {
+            videoWriter = try AVAssetWriter(url: outputFileURL, fileType: .mp4)
+        } catch {
+            SentryLog.error("[Session Replay] Failed to create video writer, reason: \(error)")
+            return completion(.failure(error))
         }
 
+        SentryLog.debug("[Session Replay] Creating pixel buffer based video writer input")
+        let videoWriterInput = AVAssetWriterInput(mediaType: .video, outputSettings: createVideoSettings(width: videoWidth, height: videoHeight))
+        guard let currentPixelBuffer = SentryPixelBuffer(size: pixelSize, videoWriterInput: videoWriterInput) else {
+            SentryLog.error("[Session Replay] Failed to render video, reason: pixel buffer creation failed")
+            return completion(.failure(SentryOnDemandReplayError.cantCreatePixelBuffer))
+        }
         videoWriter.add(videoWriterInput)
         videoWriter.startWriting()
         videoWriter.startSession(atSourceTime: .zero)
-        
+
         var lastImageSize: CGSize = image.size
         var usedFrames = [SentryReplayFrame]()
-        let group = DispatchGroup()
+        var frameIndex = from
+
+        // Convenience wrapper to handle the completion callback to return the video info and the final frame index
+        // It is not possible to use an inout frame index here, because the closure is escaping and the frameIndex variable is captured.
+        let deferredCompletionCallback: (Result<SentryVideoInfo?, Error>) -> Void = { result in
+            switch result {
+            case .success(let videoResult):
+                completion(.success(SentryRenderVideoResult(
+                    info: videoResult,
+                    finalFrameIndex: frameIndex
+                )))
+            case .failure(let error):
+                completion(.failure(error))
+            }
+        }
         
-        var result: Result<SentryVideoInfo?, Error>?
-        var frameCount = from
-        
-        group.enter()
-        videoWriterInput.requestMediaDataWhenReady(on: workingQueue.queue) {
+        // Append frames to the video writer input in a pull-style manner when the input is ready to receive more media data.
+        //
+        // Inside the callback:
+        // 1. We append media data until `isReadyForMoreMediaData` becomes false
+        // 2. Or until there's no more media data to process (then we mark input as finished)
+        // 3. If we don't mark the input as finished, the callback will be invoked again
+        //    when the input is ready for more data
+        //
+        // By setting the queue to the asset worker queue, we ensure that the callback is invoked on the asset worker queue.
+        // This is important to avoid a deadlock, as this method is called on the processing queue.
+        videoWriterInput.requestMediaDataWhenReady(on: assetWorkerQueue.queue) { [weak self] in
             SentryLog.debug("[Session Replay] Video writer input is ready, status: \(videoWriter.status)")
+            guard let strongSelf = self else {
+                SentryLog.warning("[Session Replay] On-demand replay is deallocated, completing writing session without output video info")
+                return deferredCompletionCallback(.success(nil))
+            }
             guard videoWriter.status == .writing else {
-                SentryLog.error("[Session Replay] Video writer status is not writing, cancelling video writing")
+                SentryLog.error("[Session Replay] Video writer is not writing anymore, cancelling the writing session, reason: \(videoWriter.error?.localizedDescription ?? "Unknown error")")
                 videoWriter.cancelWriting()
-                result = .failure(videoWriter.error ?? SentryOnDemandReplayError.errorRenderingVideo )
-                group.leave()
-                return
+                return deferredCompletionCallback(.failure(videoWriter.error ?? SentryOnDemandReplayError.errorRenderingVideo))
             }
-            if frameCount >= videoFrames.count {
-                SentryLog.debug("[Session Replay] Frame count is greater than video frames count, finishing video")
-                result = self.finishVideo(outputFileURL: outputFileURL, usedFrames: usedFrames, videoHeight: Int(videoHeight), videoWidth: Int(videoWidth), videoWriter: videoWriter)
-                group.leave()
-                return
+            guard frameIndex < videoFrames.count else {
+                SentryLog.debug("[Session Replay] No more frames available to process, finishing the video")
+                return strongSelf.finishVideo(
+                    outputFileURL: outputFileURL,
+                    usedFrames: usedFrames,
+                    videoHeight: Int(videoHeight),
+                    videoWidth: Int(videoWidth),
+                    videoWriter: videoWriter,
+                    onCompletion: deferredCompletionCallback
+                )
             }
-            let frame = videoFrames[frameCount]
+
+            let frame = videoFrames[frameIndex]
             if let image = UIImage(contentsOfFile: frame.imagePath) {
-                SentryLog.debug("[Session Replay] Image is ready, size: \(image.size)")
-                if lastImageSize != image.size {
+                SentryLog.debug("[Session Replay] Image at index \(frameIndex) is ready, size: \(image.size)")
+                guard lastImageSize == image.size else {
                     SentryLog.debug("[Session Replay] Image size has changed, finishing video")
-                    result = self.finishVideo(outputFileURL: outputFileURL, usedFrames: usedFrames, videoHeight: Int(videoHeight), videoWidth: Int(videoWidth), videoWriter: videoWriter)
-                    group.leave()
-                    return
+                    return strongSelf.finishVideo(
+                        outputFileURL: outputFileURL,
+                        usedFrames: usedFrames,
+                        videoHeight: Int(videoHeight),
+                        videoWidth: Int(videoWidth),
+                        videoWriter: videoWriter,
+                        onCompletion: deferredCompletionCallback
+                    )
                 }
                 lastImageSize = image.size
-                
+
                 let presentTime = SentryOnDemandReplay.calculatePresentationTime(
-                    forFrameAtIndex: frameCount,
-                    frameRate: self.frameRate
+                    forFrameAtIndex: frameIndex,
+                    frameRate: strongSelf.frameRate
                 ).timeValue
-                if currentPixelBuffer.append(image: image, presentationTime: presentTime) != true {
+                guard currentPixelBuffer.append(image: image, presentationTime: presentTime) else {
+                    SentryLog.error("[Session Replay] Failed to append image to pixel buffer, cancelling the writing session, reason: \(String(describing: videoWriter.error))")
                     videoWriter.cancelWriting()
-                    result = .failure(videoWriter.error ?? SentryOnDemandReplayError.errorRenderingVideo )
-                    group.leave()
-                    return
+                    return deferredCompletionCallback(.failure(videoWriter.error ?? SentryOnDemandReplayError.errorRenderingVideo))
                 }
                 usedFrames.append(frame)
             }
-            frameCount += 1
+
+            // Increment the frame index even if the image could not be appended to the pixel buffer.
+            // This is important to avoid an infinite loop.
+            frameIndex += 1
         }
-        guard group.wait(timeout: .now() + 2) == .success else { throw SentryOnDemandReplayError.errorRenderingVideo }
-        from = frameCount
-        
-        return try result?.get()
     }
-        
-    private func finishVideo(outputFileURL: URL, usedFrames: [SentryReplayFrame], videoHeight: Int, videoWidth: Int, videoWriter: AVAssetWriter) -> Result<SentryVideoInfo?, Error> {
+
+    // swiftlint:enable function_body_length cyclomatic_complexity
+    private func finishVideo(
+        outputFileURL: URL,
+        usedFrames: [SentryReplayFrame],
+        videoHeight: Int,
+        videoWidth: Int,
+        videoWriter: AVAssetWriter,
+        onCompletion completion: @escaping (Result<SentryVideoInfo?, Error>) -> Void
+    ) {
+        // Note: This method is expected to be called from the asset worker queue and *not* the processing queue.
         SentryLog.info("[Session Replay] Finishing video with output file URL: \(outputFileURL), used frames count: \(usedFrames.count), video height: \(videoHeight), video width: \(videoWidth)")
-        let group = DispatchGroup()
-        var finishError: Error?
-        var result: SentryVideoInfo?
-        
-        group.enter()
         videoWriter.inputs.forEach { $0.markAsFinished() }
         videoWriter.finishWriting { [weak self] in
-            defer { group.leave() }
-
             SentryLog.debug("[Session Replay] Finished video writing, status: \(videoWriter.status)")
             guard let strongSelf = self else {
                 SentryLog.warning("[Session Replay] On-demand replay is deallocated, completing writing session without output video info")
-                return
+                return completion(.success(nil))
             }
 
             switch videoWriter.status {
             case .writing:
                 SentryLog.error("[Session Replay] Finish writing video was called with status writing, this is unexpected! Completing with no video info")
+                completion(.success(nil))
             case .cancelled:
                 SentryLog.warning("[Session Replay] Finish writing video was cancelled, completing with no video info.")
+                completion(.success(nil))
             case .completed:
                 SentryLog.debug("[Session Replay] Finish writing video was completed, creating video info from file attributes.")
                 do {
-                    result = try strongSelf.getVideoInfo(
+                    let result = try strongSelf.getVideoInfo(
                         from: outputFileURL,
                         usedFrames: usedFrames,
                         videoWidth: Int(videoWidth),
                         videoHeight: Int(videoHeight)
                     )
+                    completion(.success(result))
                 } catch {
-                    SentryLog.warning("[Session Replay] Failed to create video info from file attributes, reason: \(error.localizedDescription)")
-                    finishError = error
+                    SentryLog.warning("[Session Replay] Failed to create video info from file attributes, reason: \(error)")
+                    completion(.failure(error))
                 }
-            case .failed, .unknown:
-                SentryLog.warning("[Session Replay] Finish writing video failed, reason: \(videoWriter.error?.localizedDescription ?? "Unknown error")")
-                finishError = videoWriter.error ?? SentryOnDemandReplayError.errorRenderingVideo
+            case .failed:
+                SentryLog.warning("[Session Replay] Finish writing video failed, reason: \(String(describing: videoWriter.error))")
+                completion(.failure(videoWriter.error ?? SentryOnDemandReplayError.errorRenderingVideo))
+            case .unknown:
+                SentryLog.warning("[Session Replay] Finish writing video with unknown status, reason: \(String(describing: videoWriter.error))")
+                completion(.failure(videoWriter.error ?? SentryOnDemandReplayError.errorRenderingVideo))
             @unknown default:
-                SentryLog.warning("[Session Replay] Finish writing video failed, reason: \(videoWriter.error?.localizedDescription ?? "Unknown error")")
-                finishError = videoWriter.error ?? SentryOnDemandReplayError.errorRenderingVideo
+                SentryLog.warning("[Session Replay] Finish writing video in unknown state, reason: \(String(describing: videoWriter.error))")
+                completion(.failure(SentryOnDemandReplayError.errorRenderingVideo))
             }
         }
-        group.wait()
-        
-        if let finishError = finishError { return .failure(finishError) }
-        return .success(result)
-    }
-    
-    private func filterFrames(beginning: Date, end: Date) -> [SentryReplayFrame] {
-        var frames = [SentryReplayFrame]()
-        // Using dispatch queue as sync mechanism since we need a queue already to generate the video.
-        workingQueue.dispatchSync {
-            frames = self._frames.filter { $0.time >= beginning && $0.time <= end }
-        }
-        return frames
     }
 
     fileprivate func getVideoInfo(from outputFileURL: URL, usedFrames: [SentryReplayFrame], videoWidth: Int, videoHeight: Int) throws -> SentryVideoInfo {
@@ -409,4 +520,4 @@ class SentryOnDemandReplay: NSObject, SentryReplayVideoMaker {
 
 #endif // os(iOS) || os(tvOS)
 #endif // canImport(UIKit)
-// swiftlint:enable file_length
+// swiftlint:enable file_length type_body_length